Rewrite the storage driver into python.
Now that all repositories use the same sync scripts, it's time to overhaul
the driver scripts.  This converts to a single systemd service, with different
instances for each repository.  The differences between each repository are
now encoded in configuration files instead of a bash script per repo.

Needs testing, but it's on the right track.